#1 feat : 루트 페이지를 참가 화면, main 페이지 라우팅 구현
